Bug 1705281 - Part 1: Accommodate exceptions in `deleteAllTasks`; don't try to delete WDBA task. r=agashlin
It makes sense to accommodate exceptions in `deleteAllTasks`: since
we're trying to clean up, do as much as possible.  It's possible that
we should catch, store, and re-throw any exception rather than trying
to delete a second time, but it's not clear how that re-thrown
exception will present in the browser console, etc.

We special case the WDBA task in the task scheduler implementation
rather than having consumers accommodate it because the task is the
outlier that behaves differently from the rest of the task scheduler,
and because we don't anticipate additional outlier tasks in the
future.

Depends on D110784

Differential Revision: https://phabricator.services.mozilla.com/D111947

UltraBlame original commit: 62bd2c72a28814dd45fe95b77fa5cf680816a3a3

+++ toolkit/components/taskscheduler/TaskSchedulerWinImpl.jsm
@@ -11,6 +11,9 @@ const { XPCOMUtils } = ChromeUtils.import(
   "resource://gre/modules/XPCOMUtils.jsm"
 );
 
+const { AppConstants } = ChromeUtils.import(
+  "resource://gre/modules/AppConstants.jsm"
+);
 const { Services } = ChromeUtils.import("resource://gre/modules/Services.jsm");
 
 XPCOMUtils.defineLazyServiceGetters(this, {
@@ -57,6 +60,14 @@ var _TaskSchedulerWinImpl = {
     WinTaskSvc.deleteTask(this._taskFolderName(), this._formatTaskName(id));
   },
 
+  
+
+
+
+
+
+
+
   deleteAllTasks() {
     const taskFolderName = this._taskFolderName();
 
@@ -73,11 +84,36 @@ var _TaskSchedulerWinImpl = {
 
     const tasksToDelete = allTasks.filter(name => this._matchAppTaskName(name));
 
+    let numberDeleted = 0;
+    let lastFailedTaskName;
+    
+    const defaultBrowserAgentTaskName =
+      AppConstants.MOZ_APP_DISPLAYNAME_DO_NOT_USE +
+      " Default Browser Agent " +
+      XreDirProvider.getInstallHash();
     for (const taskName of tasksToDelete) {
-      WinTaskSvc.deleteTask(taskFolderName, taskName);
+      if (taskName == defaultBrowserAgentTaskName) {
+        
+        continue;
+      }
+
+      try {
+        WinTaskSvc.deleteTask(taskFolderName, taskName);
+        numberDeleted += 1;
+      } catch (e) {
+        lastFailedTaskName = taskName;
+      }
     }
 
-    if (allTasks.length == tasksToDelete.length) {
+    if (lastFailedTaskName) {
+      
+      
+      
+      
+      WinTaskSvc.deleteTask(taskFolderName, lastFailedTaskName);
+    }
+
+    if (allTasks.length == numberDeleted) {
       
       this._deleteFolderIfEmpty();
     }
